move comm open handler / comm target exception over to client.py
